@@ -6,20 +6,14 @@ import (
 	"fmt"
 	"github.com/aws/aws-lambda-go/events"
 	"github.com/aws/aws-lambda-go/lambda"
-	errExt "github.com/shekhar-jha/base-demo/go-utils/pkg/err"
-	"github.com/shekhar-jha/base-demo/go-utils/pkg/log"
+	logger "github.com/grinps/go-utils/base-utils/logs"
+	"github.com/grinps/go-utils/errext"
 	"net/http"
 	"os"
 	"strings"
 )
 
-func init() {
-	log.Configure(log.NewConfigLogger(log.Debug, "GoLogger"), "github.com/shekhar-jha/base-demo")
-}
-
-var logger = log.GetLogger("github.com/shekhar-jha/base-demo", "go-lambda", "cmd")
-
-type Handle[Request any, Response any] func(context Context[Request, Response]) errExt.Error
+type Handle[Request any, Response any] func(context Context[Request, Response]) error
 
 type Context[Request any, Response any] interface {
 	GetRequest() *Request
@@ -41,7 +35,7 @@ func (context *simpleContext[Request, Response]) SetResponse(response Response) 
 
 type Application[Request any, Response any] interface {
 	RegisterHandler(handle Handle[Request, Response]) Application[Request, Response]
-	Run() errExt.Error
+	Run() error
 }
 
 func NewApplication[Request any, Response any]() Application[Request, Response] {
@@ -58,7 +52,7 @@ type httpHandler[Request any, Response any] struct {
 }
 
 func (handler *httpHandler[Request, Response]) ServeHTTP(writer http.ResponseWriter, request *http.Request) {
-	logger.LogWithLevel(log.Debug, "Invoked request with %v", request)
+	logger.Log("Invoked request with %v", request)
 	var requestObjectMap = map[string]interface{}{}
 	for key, value := range request.URL.Query() {
 		if len(value) >= 1 {
@@ -71,9 +65,9 @@ func (handler *httpHandler[Request, Response]) ServeHTTP(writer http.ResponseWri
 		return
 	}
 	var simpleContext = &simpleContext[Request, Response]{request: &requestObject}
-	logger.LogWithLevel(log.Debug, "Invoking handler with request %#v", simpleContext.request)
+	logger.Log("Invoking handler with request", simpleContext.request)
 	err := handler.handler(simpleContext)
-	logger.LogWithLevel(log.Debug, "Response %#v; Error %v", simpleContext.response, err)
+	logger.Log("Response %#v; Error", simpleContext.response, err)
 	if err != nil {
 		writeError(writer, http.StatusInternalServerError, err)
 	} else {
@@ -85,29 +79,29 @@ func (handler *httpHandler[Request, Response]) ServeHTTP(writer http.ResponseWri
 			writer.Header().Set("Content-Type", "application/json")
 			size, writeErr := writer.Write(responseOut)
 			if writeErr != nil {
-				logger.LogWithLevel(log.Warn, "Failed to write error response %v (only %d written) due to error %v", simpleContext.response, size, writeErr)
+				logger.Warn("Failed to write error response", simpleContext.response, "(only", size, "written) due to error ", writeErr)
 			}
 		}
 	}
 }
 
-func writeError(writer http.ResponseWriter, code int, err errExt.Error) {
+func writeError(writer http.ResponseWriter, code int, err error) {
 	writer.WriteHeader(code)
 	writer.Header().Set("Content-Type", "application/json")
 	size, writeErr := writer.Write([]byte("{ \"Response\" : \"" + err.Error() + "\"}"))
 	if writeErr != nil {
-		logger.LogWithLevel(log.Warn, "Failed to write response %v (only %d written) due to error %v", err, size, writeErr)
+		logger.Warn("Failed to write response", err, "(only ", size, "written) due to error", writeErr)
 	}
 }
 
 func (app *simpleApplication[Request, Response]) RegisterHandler(handler Handle[Request, Response]) Application[Request, Response] {
 	if os.Getenv("AWS_LAMBDA_RUNTIME_API") != "" {
-		logger.LogWithLevel(log.Info, "AWS Lambda runtime detected")
+		logger.Log("AWS Lambda runtime detected")
 		app.lambdaHandler = lambda.NewHandler(lambdaHandler(func(ctx context.Context, bytes []byte) ([]byte, error) {
 			return simpleLambdaHandler(handler, ctx, bytes)
 		}))
 	} else if os.Getenv("PORT") != "" {
-		logger.LogWithLevel(log.Info, "GCP Cloud Run runtime detected.")
+		logger.Log("GCP Cloud Run runtime detected.")
 		app.httpHandler = http.NewServeMux()
 		app.httpHandler.Handle("/", &httpHandler[Request, Response]{
 			handler: handler,
@@ -116,15 +110,15 @@ func (app *simpleApplication[Request, Response]) RegisterHandler(handler Handle[
 	return app
 }
 
-var ErrStartupFailed = errExt.NewErrorTemplate("Server Start Failed", "Failed to start server due to error {{ .Error }}")
+var ErrStartupFailed = errext.NewErrorCodeWithOptions(errext.WithTemplate("Failed to start server due to error", "[error]"))
 
-func (app *simpleApplication[Request, Response]) Run() errExt.Error {
+func (app *simpleApplication[Request, Response]) Run() error {
 	if app.lambdaHandler != nil {
 		lambda.Start(app.lambdaHandler)
 	} else {
 		err := http.ListenAndServe(":"+os.Getenv("PORT"), app.httpHandler)
 		if err != nil {
-			return ErrStartupFailed.NewWithError(err)
+			return ErrStartupFailed.NewWithErrorF(err, "error", err)
 		}
 	}
 	return nil
@@ -139,54 +133,48 @@ func (handler lambdaHandler) Invoke(ctx context.Context, payload []byte) ([]byte
 type ProtocolType int
 
 const (
-	Unknown ProtocolType = iota
+	ProtocolTypeUnknown ProtocolType = iota
 	//	Support various protocols like TCP, HTTP, etc
 	AwsLambdaHttp
 	AwsLambdaInvoke
 )
 
-var ErrRequestParsingError = errExt.NewErrorTemplate("Parsing Error", "Failed to parse {{ .Data.Parameter }} payload as JSON due to error {{ .Data.ParseError }}")
-var ErrResponseGenerationError = errExt.NewErrorTemplate("Response Generation Error", "Failed to generate response payload due to error {{ .Data }}")
+var ErrRequestParsingError = errext.NewErrorCodeWithOptions(errext.WithTemplate("Failed to parse", "[Parameter]", "payload as JSON due to error", "[ParseError]"))
+var ErrResponseGenerationError = errext.NewErrorCodeWithOptions(errext.WithTemplate("Failed to generate response payload due to error", "[error]"))
 
 func simpleLambdaHandler[Request any, Response any](handler Handle[Request, Response], ctx context.Context, payload []byte) ([]byte, error) {
 	readData := map[string]interface{}{}
 	unmarshalErr := json.Unmarshal(payload, &readData)
 	if unmarshalErr != nil {
-		return nil, ErrRequestParsingError.New(struct {
-			Parameter  string
-			ParseError error
-		}{Parameter: "Request", ParseError: unmarshalErr})
+		return nil, ErrRequestParsingError.NewWithErrorF(unmarshalErr, "Parameter", "Request", "ParseError", unmarshalErr)
 	}
 	requestObject := map[string]interface{}{}
-	logger.LogWithLevel(log.Debug, "Parsed request object %#v", readData)
+	logger.Log("Parsed request object", readData)
 	var eventType ProtocolType
 	if _, isHTTP := asMap(readData, "requestContext", "http"); isHTTP {
-		logger.LogWithLevel(log.Debug, "Request is an Lambda HTTP Request")
+		logger.Log("Request is an Lambda HTTP Request")
 		eventType = AwsLambdaHttp
 		requestId, _ := asString(readData, "requestContext", "requestId")
-		logger.LogWithLevel(log.Debug, "Request Id %s", requestId)
+		logger.Log("Request Id", requestId)
 		if queryParams, hasQueryParams := asMap(readData, "queryStringParameters"); hasQueryParams {
 			copyMap(queryParams, requestObject)
-			logger.LogWithLevel(log.Debug, "Added query parameters %#v", queryParams)
+			logger.Log("Added query parameters", queryParams)
 		}
 		if body, hasBody := asString(readData, "body"); hasBody {
-			logger.LogWithLevel(log.Debug, "Request body is present")
+			logger.Log("Request body is present")
 			if contentType, hasContentType := asString(readData, "headers", "content-type"); hasContentType {
 				switch strings.ToLower(contentType) {
 				case "application/json":
-					logger.LogWithLevel(log.Debug, "Parsing content type application/json")
+					logger.Log("Parsing content type application/json")
 					bodyUnmarshalError := json.Unmarshal([]byte(body), &requestObject)
 					if bodyUnmarshalError != nil {
-						return nil, ErrRequestParsingError.New(struct {
-							Parameter  string
-							ParseError error
-						}{Parameter: "Body", ParseError: bodyUnmarshalError})
+						return nil, ErrRequestParsingError.NewWithErrorF(unmarshalErr, "Parameter", "Body", "ParseError", bodyUnmarshalError)
 					}
 				default:
-					logger.LogWithLevel(log.Warn, "Failed to identify request body content type %s for request %s", contentType, requestId)
+					logger.Warn("Failed to identify request body content type", contentType, "for request", requestId)
 				}
 			} else {
-				logger.LogWithLevel(log.Warn, "Request body is present but no content-type header is available for validation.")
+				logger.Warn("Request body is present but no content-type header is available for validation.")
 			}
 		}
 	} else {
@@ -199,9 +187,9 @@ func simpleLambdaHandler[Request any, Response any](handler Handle[Request, Resp
 	}
 	var requestContext = &simpleContext[Request, Response]{request: &requestObjectAsStruct}
 
-	logger.LogWithLevel(log.Debug, "Invoking handler with context %v, request %#v", ctx, requestContext.request)
+	logger.Log("Invoking handler with context", ctx, "request", requestContext.request)
 	err := handler(requestContext)
-	logger.LogWithLevel(log.Debug, "Response %#v; Error %v", requestContext.response, err)
+	logger.Log("Response", requestContext.response, "Error", err)
 	switch eventType {
 	case AwsLambdaInvoke:
 		if err != nil {
@@ -232,7 +220,7 @@ func simpleLambdaHandler[Request any, Response any](handler Handle[Request, Resp
 		}
 		respByte, httpMarshalErr := json.Marshal(response)
 		return respByte, httpMarshalErr
-	case Unknown:
+	case ProtocolTypeUnknown:
 		unsupportedFormatErr := fmt.Sprintf("Event type has not been set")
 		return nil, ErrResponseGenerationError.New(unsupportedFormatErr)
 	default:
@@ -262,19 +250,18 @@ func asString(input map[string]interface{}, key ...string) (string, bool) {
 	return "", false
 }
 func asMap(input map[string]interface{}, key ...string) (map[string]interface{}, bool) {
-	asMapLogger := logger.Push("asMap", key)
-	logger.LogWithLevel(log.Trace, "Input: %#v", input)
+	logger.Log("Input:", input)
 	returnValue := map[string]interface{}{}
 	if input == nil {
 		return returnValue, false
 	}
 	totalKeys := len(key)
-	asMapLogger.LogWithLevel(log.Trace, "Total keys: %d", totalKeys)
+	logger.Log("Total keys:", totalKeys)
 	if totalKeys > 0 {
 		var applicableMap = input
 		for remainingKeys := totalKeys; remainingKeys >= 1; remainingKeys-- {
-			asMapLogger.LogWithLevel(log.Trace, "Applicable map: %v", applicableMap)
-			asMapLogger.LogWithLevel(log.Trace, "Remaining keys: %d", remainingKeys)
+			logger.Log("Applicable map:", applicableMap)
+			logger.Log("Remaining keys:", remainingKeys)
 			if value, hasValue := applicableMap[key[totalKeys-remainingKeys]]; hasValue && value != nil {
 				if valueAsMap, isMap := value.(map[string]interface{}); isMap {
 					applicableMap = valueAsMap
@@ -297,21 +284,17 @@ func copyMap(src map[string]interface{}, dest map[string]interface{}) {
 	}
 }
 
-func createObject[Request any](requestObject map[string]interface{}) (Request, errExt.Error) {
+func createObject[Request any](requestObject map[string]interface{}) (Request, error) {
 	var requestObjectAsStruct Request
 	generatedRequest, requestMarshalErr := json.Marshal(requestObject)
 	if requestMarshalErr != nil {
-		return requestObjectAsStruct, ErrRequestParsingError.New(struct {
-			Parameter  string
-			ParseError error
-		}{Parameter: "ObjectMarshal", ParseError: requestMarshalErr})
+		return requestObjectAsStruct, ErrRequestParsingError.NewWithErrorF(requestMarshalErr,
+			"Parameter", "ObjectMarshal", "ParseError", requestMarshalErr)
 	}
 	requestUnmarshalErr := json.Unmarshal(generatedRequest, &requestObjectAsStruct)
 	if requestUnmarshalErr != nil {
-		return requestObjectAsStruct, ErrRequestParsingError.New(struct {
-			Parameter  string
-			ParseError error
-		}{Parameter: "Request Object", ParseError: requestUnmarshalErr})
+		return requestObjectAsStruct, ErrRequestParsingError.NewWithErrorF(requestMarshalErr,
+			"Request Object", "ObjectMarshal", "ParseError", requestUnmarshalErr)
 	}
 	return requestObjectAsStruct, nil
 }
